refactor: rename ambiguous Setup Wizard Operator base classes file
