+++ setup_wizard/fix_mouth_outlines.py
@@ -4,7 +4,7 @@ import bpy
 from bpy.types import Operator
 
 from setup_wizard.import_order import NextStepInvoker
-from setup_wizard.models import CustomOperatorProperties
+from setup_wizard.setup_wizard_operator_base_classes import CustomOperatorProperties
 
 CAMERA_INPUT = 'Input_4'
 DEPTH_OFFSET_INPUT = 'Input_8'
