+++ setup_wizard/genshin_import_materials.py
@@ -11,7 +11,7 @@ from setup_wizard.import_order import NextStepInvoker
 
 from setup_wizard.material_import_setup.game_material_importers import GameMaterialImporterFactory
 from setup_wizard.material_import_setup.material_importer_service import MaterialImporterService
-from setup_wizard.models import BasicSetupUIOperator, CustomOperatorProperties
+from setup_wizard.setup_wizard_operator_base_classes import BasicSetupUIOperator, CustomOperatorProperties
 
 
 class GI_OT_SetUpMaterials(Operator, BasicSetupUIOperator):
