+++ setup_wizard/genshin_import_outline_lightmaps.py
@@ -8,7 +8,7 @@ from bpy_extras.io_utils import ImportHelper
 from bpy.props import StringProperty
 from bpy.types import Operator
 from setup_wizard.import_order import NextStepInvoker
-from setup_wizard.models import CustomOperatorProperties
+from setup_wizard.setup_wizard_operator_base_classes import CustomOperatorProperties
 from setup_wizard.texture_import_setup.outline_texture_importers import OutlineTextureImporter, OutlineTextureImporterFactory
