+++ setup_wizard/misc_operations.py
@@ -4,7 +4,7 @@ import bpy
 from bpy.types import Operator
 
 from setup_wizard.import_order import NextStepInvoker
-from setup_wizard.models import CustomOperatorProperties
+from setup_wizard.setup_wizard_operator_base_classes import CustomOperatorProperties
 
 
 class GI_OT_SetColorManagementToStandard(Operator, CustomOperatorProperties):
