+++ setup_wizard/set_up_head_driver.py
@@ -4,7 +4,7 @@ import bpy
 from bpy.types import Operator
 
 from setup_wizard.import_order import NextStepInvoker
-from setup_wizard.models import CustomOperatorProperties
+from setup_wizard.setup_wizard_operator_base_classes import CustomOperatorProperties
 
 HEAD_DRIVER_OBJECT_NAME = 'Head Driver'
